Properly change PATH session variable.
The command now properly changes the PATH variable in the current shell
session as well as the ~/.profile for future sessions.

There was some confusion caused byt the fact that script needs to be
executed using the 'source' command in order to be able to change the
PATH variable in the terminal session that launched the script.